Refacto code by chunking task/index.js into multiple files and have a single context at task load
