+++ frontend/index.tsx
@@ -19,6 +19,7 @@ import {ActionTypes as PlayerActionTypes} from "./player/actionTypes";
 import {ActionTypes as EditorActionTypes} from "./editor/actionTypes";
 import {ActionTypes as RecorderActionTypes} from "./recorder/actionTypes";
 import {ActionTypes as StatisticsActionTypes} from "./statistics/actionTypes";
+import {ActionTypes as TaskActionTypes} from "./task/actionTypes";
 import {SandboxApp} from "./sandbox/SandboxApp";
 import {TaskApp} from "./task/TaskApp";
 import {StatisticsApp} from "./statistics/StatisticsApp";
@@ -221,6 +222,10 @@ Codecast.start = function(options) {
             break;
 
         case 'task':
+            store.dispatch({
+                type: TaskActionTypes.TaskLoad,
+            });
+
             appDisplay = <TaskApp />;
 
             break;
