+++ frontend/task/actionTypes.ts
@@ -1,3 +1,4 @@
 export enum ActionTypes {
   TaskLoad = 'Task.Load',
+  TaskUpdateContext = 'Task.UpdateContext',
 }
